Move fetchVocabulary to customerApi, export only one vocab-translation
