+++ src/pages/editor/VocabularyRow.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { CustomerVocabulary, VocabularyStatus } from '../../types/customerInstitution.types';
 import { dataTestId as dataTestIds } from '../../utils/dataTestIds';
-import { getTranslatedVocabularyName } from './VocabularySettings';
+import { getTranslatedVocabularyName } from './VocabularyOverview';
 
 const toggleButtonSx: SxProps = { width: 'fit-content', px: '1rem' };
